Cleaning up and re-organizing Synthetix tests
